ref admin panel work at products

+++ database/migrations/2021_05_11_072499_create_margin.php
@@ -16,7 +16,7 @@ class CreateMargin extends Migration
         Schema::create('margin', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name')->nullable();
-            $table->integer('value')->default(0);
+            $table->decimal('value',8,2)->default(0);
             $table->timestamps();
         });
     }
