+++ database/migrations/2021_05_11_072500_create_discount_table.php
@@ -16,7 +16,7 @@ class CreateDiscountTable extends Migration
         Schema::create('discount', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name')->nullable();
-            $table->integer('value')->default(0);
+            $table->decimal('value',8,2)->default(0);
             $table->timestamps();
         });
     }
